Refactor: Restructure table column imports and remove unused files
- Updated import paths for order, payment type, and payment method columns to reflect new directory structure.
- Deleted obsolete files for order columns, payment type columns, payment type group columns, and payment method columns to clean up the codebase.
